LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-opener-policy/reporting/navigation-reporting/reporting-redirect-with-same-origin-allow-popups.https-expected.txt


FAIL Same origin openee redirected to same-origin with same-origin-allow-popups assert_equals: expected (string) "coop" but got (undefined) undefined
FAIL Cross origin openee redirected to same-origin with same-origin-allow-popups assert_equals: expected (string) "coop" but got (undefined) undefined
